# docs/swagger.yaml
basePath: /api
definitions:
  Attachment:
    properties:
      filename:
        type: string
      filetype:
        type: string
      url:
        type: string
    type: object
...
  ErrorResponse:
    properties:
      error:
        $ref: '#/definitions/HttpError'
    type: object
  ErrorsResponse:
    properties:
      errors:
        items:
          $ref: '#/definitions/FieldError'
        type: array
    type: object
  FieldError:
    properties:
      field:
        description: The property containing the error
        type: string
      message:
        description: The specific error message
        type: string
    type: object
...
  HttpError:
    properties:
      message:
        description: The specific error message
        type: string
      type:
        description: The Http Response as a string
        type: string
    type: object
...
  Member:
    properties:
      color:
        type: string
      createdAt:
        type: string
      id:
        type: string
      image:
        type: string
      isFriend:
        type: boolean
      isOnline:
        type: boolean
      nickname:
        type: string
      updatedAt:
        type: string
      username:
        type: string
    type: object
...
  Message:
    properties:
      attachment:
        $ref: '#/definitions/Attachment'
      createdAt:
        type: string
      id:
        type: string
      text:
        type: string
      updatedAt:
        type: string
      user:
        $ref: '#/definitions/Member'
    type: object
  MessageRequest:
    properties:
      file:
        description: image/* or audio/*
        format: binary
        type: string
      text:
        description: Maximum 2000 characters
        type: string
    type: object
...
  SuccessResponse:
    properties:
      success:
        description: Only returns true, not a json object
        type: boolean
    type: object
...
host: localhost:<PORT>
info:
  contact: {}
  description: Xonia REST API . This service uses sessions for authentication
  license:
    name: GNUv3
  title: Xonia API
  version: "2.0.0"
paths:
...
  /messages/{channelId}:
    get:
      parameters:
      - description: Channel ID
        in: path
        name: channelId
        required: true
        type: string
      - description: Cursor Pagination using the createdAt field
        in: query
        name: cursor
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/Message'
            type: array
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/ErrorResponse'
      summary: Get Channel Messages
      tags:
      - Messages
    post:
      parameters:
      - description: Channel ID
        in: path
        name: channelId
        required: true
        type: string
      - description: Create Message
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/MessageRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/SuccessResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/ErrorsResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/ErrorResponse'
      summary: Create Messages
      tags:
      - Messages
  /messages/{messageId}:
    delete:
      parameters:
      - description: Message ID
        in: path
        name: messageId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/SuccessResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/ErrorResponse'
      summary: Delete Messages
      tags:
      - Messages
    put:
      parameters:
      - description: Message ID
        in: path
        name: messageId
        required: true
        type: string
      - description: Edit Message
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/MessageRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/SuccessResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/ErrorsResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/ErrorResponse'
      summary: Edit Messages
      tags:
      - Messages
swagger: "2.0"